Change `capacity` type from `Int` to `Double` across models and add ISO 8601 date formatting for `HistoryScreen`.

+++ composeApp/src/commonMain/kotlin/com/fruex/beerwall/data/remote/dto/history/History.kt
@@ -18,7 +18,7 @@ data class TransactionResponse(
     val commodityName: String,
     val startDateTime: String,
     val grossPrice: Double,
-    val capacity: Int
+    val capacity: Double
 )
 
 /**
